feat(*): implement authentication

+++ todo/src/item/item.service.ts
@@ -4,20 +4,23 @@ import { PrismaService } from 'nestjs-prisma'
 import { ErrorUtil } from 'src/common/utils/error.util'
 import { CreateDto } from './dto/create.dto'
 import { UpdateDto } from './dto/update.dto'
+import { ItemNotFoundException } from './exceptions'
 
 @Injectable()
 export class ItemService {
   constructor(private prisma: PrismaService) {}
 
-  async create(createDto: CreateDto): Promise<Item> {
-    // TODO: Make sure listId belongs to the user
+  async create(userId: string, createDto: CreateDto): Promise<Item> {
     try {
       const item = await this.prisma.item.create({
         data: {
           title: createDto.title,
           list: {
             connect: {
-              id: createDto.listId,
+              userIndex: {
+                userId,
+                id: createDto.listId,
+              },
             },
           },
         },
@@ -33,9 +36,21 @@ export class ItemService {
     }
   }
 
-  async delete(id: string): Promise<void> {
-    // TODO: Make sure item belongs to the user
+  async delete(userId: string, id: string): Promise<void> {
     try {
+      const item = await this.prisma.item.findUniqueOrThrow({
+        where: {
+          id,
+        },
+        include: {
+          list: true,
+        },
+      })
+
+      if (item.list.userId !== userId) {
+        throw new ItemNotFoundException()
+      }
+
       await this.prisma.item.delete({
         where: {
           id,
@@ -43,24 +58,33 @@ export class ItemService {
       })
     } catch (error) {
       if (ErrorUtil.isNotFoundError(error)) {
-        throw new BadRequestException('Item not found')
+        throw new ItemNotFoundException()
       }
 
       throw error
     }
   }
 
-  async update(id: string, updateDto: UpdateDto): Promise<Item> {
+  async update(userId: string, id: string, updateDto: UpdateDto): Promise<Item> {
     try {
       const item = await this.prisma.item.update({
-        where: { id },
+        where: {
+          id,
+        },
+        include: {
+          list: true,
+        },
         data: updateDto,
       })
+
+      if (item.list.userId !== userId) {
+        throw new ItemNotFoundException()
+      }
 
       return item
     } catch (error) {
       if (ErrorUtil.isNotFoundError(error)) {
-        throw new BadRequestException('Item not found')
+        throw new ItemNotFoundException()
       }
 
       throw error
